backend/services/ai_insights.py
from typing import List, Dict, Any, Optional
from datetime import datetime, timedelta
import openai
from fastapi import HTTPException
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import KMeans

from models.ai import (
    AIInsight,
    InsightType,
    Recommendation,
    AnalysisRequest,
    AnalysisResponse
)
from services.health import HealthService
from services.finance import FinanceService
from services.goals import GoalsService
from services.mood import MoodService
from services.habits import HabitsService
from services.journal import JournalService
from services.learning import LearningService
from services.social import SocialService
from services.mindfulness import MindfulnessService
from services.tasks import TasksService

class AIInsightsService:
    def __init__(
        self,
        health_service: HealthService,
        finance_service: FinanceService,
        goals_service: GoalsService,
        mood_service: MoodService,
        habits_service: HabitsService,
        journal_service: JournalService,
        learning_service: LearningService,
        social_service: SocialService,
        mindfulness_service: MindfulnessService,
        tasks_service: TasksService,
    ):
        self.health_service = health_service
        self.finance_service = finance_service
        self.goals_service = goals_service
        self.mood_service = mood_service
        self.habits_service = habits_service
        self.journal_service = journal_service
        self.learning_service = learning_service
        self.social_service = social_service
        self.mindfulness_service = mindfulness_service
        self.tasks_service = tasks_service
        
        # Initialize analysis tools
        self.scaler = StandardScaler()
        self.kmeans = KMeans(n_clusters=5, random_state=42)

    async def get_holistic_insights(self, user_id: str) -> List[AIInsight]:
        """Generate holistic insights across all areas of user data."""
        try:
            # Gather data from all services
            health_data = await self.health_service.get_user_health_data(user_id)
            finance_data = await self.finance_service.get_user_finance_data(user_id)
            goals_data = await self.goals_service.get_user_goals(user_id)
            mood_data = await self.mood_service.get_user_mood_data(user_id)
            habits_data = await self.habits_service.get_user_habits(user_id)
            journal_data = await self.journal_service.get_user_entries(user_id)
            learning_data = await self.learning_service.get_user_progress(user_id)
            social_data = await self.social_service.get_user_connections(user_id)
            mindfulness_data = await self.mindfulness_service.get_user_sessions(user_id)
            tasks_data = await self.tasks_service.get_user_tasks(user_id)

            # Analyze correlations and patterns
            insights = []
            
            # Health and Mood Correlation
            health_mood_insight = await self._analyze_health_mood_correlation(
                health_data, mood_data
            )
            if health_mood_insight:
                insights.append(health_mood_insight)

            # Financial Goals Progress
            finance_goals_insight = await self._analyze_finance_goals_progress(
                finance_data, goals_data
            )
            if finance_goals_insight:
                insights.append(finance_goals_insight)

            # Habits and Productivity
            habits_tasks_insight = await self._analyze_habits_productivity(
                habits_data, tasks_data
            )
            if habits_tasks_insight:
                insights.append(habits_tasks_insight)

            # Learning and Goals Alignment
            learning_goals_insight = await self._analyze_learning_goals_alignment(
                learning_data, goals_data
            )
            if learning_goals_insight:
                insights.append(learning_goals_insight)

            # Social and Mindfulness Impact
            social_mindfulness_insight = await self._analyze_social_mindfulness_impact(
                social_data, mindfulness_data
            )
            if social_mindfulness_insight:
                insights.append(social_mindfulness_insight)

            # Generate AI-powered recommendations
            recommendations = await self._generate_recommendations(insights)
            insights.extend(recommendations)

            return insights

        except Exception as e:
            raise HTTPException(
                status_code=500,
                detail=f"Error generating insights: {str(e)}"
            )

    async def analyze_data(self, request: AnalysisRequest) -> AnalysisResponse:
        """Analyze specific data points and generate targeted insights."""
        try:
            # Prepare data for analysis
            data_points = np.array(request.data_points)
            scaled_data = self.scaler.fit_transform(data_points.reshape(-1, 1))
            
            # Perform clustering
            clusters = self.kmeans.fit_predict(scaled_data)
            
            # Identify patterns and anomalies
            patterns = self._identify_patterns(data_points, clusters)
            anomalies = self._detect_anomalies(data_points)
            
            # Generate insights based on analysis
            insights = []
            
            # Pattern-based insights
            for pattern in patterns:
                insights.append(AIInsight(
                    type=InsightType.TREND,
                    title=f"Pattern Detected: {pattern['name']}",
                    description=pattern['description'],
                    recommendations=pattern['recommendations'],
                    related_areas=request.areas,
                    confidence=pattern['confidence'],
                    timestamp=datetime.now().isoformat()
                ))
            
            # Anomaly-based insights
            for anomaly in anomalies:
                insights.append(AIInsight(
                    type=InsightType.WARNING,
                    title=f"Anomaly Detected",
                    description=anomaly['description'],
                    recommendations=anomaly['recommendations'],
                    related_areas=request.areas,
                    confidence=anomaly['confidence'],
                    timestamp=datetime.now().isoformat()
                ))
            
            return AnalysisResponse(
                insights=insights,
                patterns=patterns,
                anomalies=anomalies
            )

        except Exception as e:
            raise HTTPException(
                status_code=500,
                detail=f"Error analyzing data: {str(e)}"
            )

    async def get_recommendations(self, area: str, user_id: str) -> List[Recommendation]:
        """Generate targeted recommendations for a specific area."""
        try:
            # Get relevant data for the area
            area_data = await self._get_area_data(area, user_id)
            
            # Generate recommendations using OpenAI
            prompt = self._create_recommendation_prompt(area, area_data)
            response = await openai.ChatCompletion.create(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": "You are an AI wellness assistant providing personalized recommendations."},
                    {"role": "user", "content": prompt}
                ]
            )
            
            # Process and structure the recommendations
            recommendations = self._process_ai_recommendations(response.choices[0].message.content)
            
            return recommendations

        except Exception as e:
            raise HTTPException(
                status_code=500,
                detail=f"Error generating recommendations: {str(e)}"
            )

    async def _analyze_health_mood_correlation(
        self,
        health_data: Dict[str, Any],
        mood_data: Dict[str, Any]
    ) -> Optional[AIInsight]:
        """Analyze correlation between health metrics and mood patterns."""
        # Implementation details for health-mood correlation analysis
        pass

    async def _analyze_finance_goals_progress(
        self,
        finance_data: Dict[str, Any],
        goals_data: Dict[str, Any]
    ) -> Optional[AIInsight]:
        """Analyze financial progress towards goals."""
        # Implementation details for finance-goals progress analysis
        pass

    async def _analyze_habits_productivity(
        self,
        habits_data: Dict[str, Any],
        tasks_data: Dict[str, Any]
    ) -> Optional[AIInsight]:
        """Analyze impact of habits on task productivity."""
        # Implementation details for habits-productivity analysis
        pass

    async def _analyze_learning_goals_alignment(
        self,
        learning_data: Dict[str, Any],
        goals_data: Dict[str, Any]
    ) -> Optional[AIInsight]:
        """Analyze alignment between learning progress and goals."""
        # Implementation details for learning-goals alignment analysis
        pass

    async def _analyze_social_mindfulness_impact(
        self,
        social_data: Dict[str, Any],
        mindfulness_data: Dict[str, Any]
    ) -> Optional[AIInsight]:
        """Analyze impact of social connections on mindfulness."""
        # Implementation details for social-mindfulness impact analysis
        pass

    def _identify_patterns(self, data: np.ndarray, clusters: np.ndarray) -> List[Dict[str, Any]]:
        """Identify patterns in the analyzed data."""
        patterns = []
        # Implementation details for pattern identification
        return patterns

    def _detect_anomalies(self, data: np.ndarray) -> List[Dict[str, Any]]:
        """Detect anomalies in the analyzed data."""
        anomalies = []
        # Implementation details for anomaly detection
        return anomalies

    async def _get_area_data(self, area: str, user_id: str) -> Dict[str, Any]:
        """Get relevant data for a specific area."""
        # Implementation details for retrieving area-specific data
        pass

    def _create_recommendation_prompt(self, area: str, data: Dict[str, Any]) -> str:
        """Create a prompt for AI recommendation generation."""
        # Implementation details for creating AI prompts
        pass

    def _process_ai_recommendations(self, ai_response: str) -> List[Recommendation]:
        """Process and structure AI-generated recommendations."""
        # Implementation details for processing AI recommendations
        pass
