backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.4.1
aiofiles==23.2.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
openai==1.3.5
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
en-core-web-sm==3.7.1
tensorflow==2.15.0
python-dotenv==1.0.0
websockets==12.0
pandas==2.1.0
numpy==1.26.2
APScheduler==3.10.1
plotly==5.17.0
pdfkit==1.0.0
jinja2==3.1.2
plaid-python==14.0.0
yodlee-api==1.0.3
nordigen-python==1.2.3
coinbase==2.1.0
cryptography==41.0.7
firebase-admin==6.2.0
fhirclient==3.2.0
fitbit==0.3.1
stravalib==1.0.0
healthkit==1.0.0
